+++ packages/core/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 0.6.11 - not officially released yet
 
 * Separate CHANGELOG from README.
+* Move the reference util functions (`referenceValues`, `incomingReferences`, `referencesToOutOfScopeNodes`) and type (class `ReferenceValue`) from the `utilities` package to `core`.
+    Aliases (as re-exports) are left in its place to avoid code breaking.
+    The move is necessary because not all clients of `@lionweb/utilities` can install the Node.js-specific dependencies of the dependencies that package (such as `crypto`).
 
 
 ## 0.6.10